+++ CHANGELOG.md
@@ -7,6 +7,7 @@ unreleased
 * Improved unicode handling in Python2 (#78)
 * Add `tsv` output format for **shodan host** (#65)
 * Show user-friendly error messages when running **shodan radar** without permission or in a window that's too small (#74)
+* Improved exception handling to improve debugging **shodan init** (#77)
 
 1.9.0
 -----